llm/requirements.txt
openai==1.31.0
virtualenv==20.23.0
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers
python-dotenv==1.0.0
langchain_community==0.2.1